First batch of settings files
Separating ACE_Settings.hpp from config.cpp
