+++ addons/interaction/config.cpp
@@ -15,10 +15,4 @@ class CfgPatches {
 #include "CfgEventHandlers.hpp"
 #include "CfgVehicles.hpp"
 #include "Menu_Config.hpp"
-
-class ACE_Settings {
-    class GVAR(EnableTeamManagement) {
-        value = 1;
-        typeName = "BOOL";
-    };
-};
+#include "ACE_Settings.hpp"
