+++ addons/inventory/ACE_Settings.hpp
@@ -1,0 +1,10 @@
+class ACE_Settings {
+    class GVAR(inventoryDisplaySize) {
+        value = 0;
+        typeName = "SCALAR";
+        isClientSettable = 1;
+        displayName = "$STR_ACE_Inventory_SettingName";
+        description = "$STR_ACE_Inventory_SettingDescription";
+        values[] = {"$str_medium", "$str_large", "$str_very_large"};
+    };
+};
