+++ addons/vehiclelock/config.cpp
@@ -12,21 +12,7 @@ class CfgPatches {
   };
 };
 
-class ACE_Settings {
-    class GVAR(DefaultLockpickStrength) {
-        value = 10;
-        typeName = "SCALAR";
-    };
-    class GVAR(LockVehicleInventory) {
-        value = 0;
-        typeName = "BOOL";
-    };    
-    class GVAR(VehicleStartingLockState) {
-        value = -1;
-        typeName = "SCALAR";
-    };
-};
-
+#include "ACE_Settings.hpp"
 #include "CfgEventHandlers.hpp"
 #include "CfgMagazines.hpp"
 #include "CfgVehicles.hpp"
